+++ src/hooks/useQueryConfig.tsx
@@ -1,0 +1,27 @@
+import { isUndefined, omitBy } from 'lodash'
+import useQueryParams from './useQueryParams'
+import { ProductListConfig } from 'src/types/product.type'
+
+export type QueryConfig = {
+  [key in keyof ProductListConfig]: string
+}
+
+export default function useQueryConfig() {
+  const queryParams = useQueryParams()
+  const queryConfig: QueryConfig = omitBy(
+    {
+      page: queryParams.page || '1',
+      limit: queryParams.limit || '5',
+      exclude: queryParams.exclude,
+      name: queryParams.name,
+      order: queryParams.order,
+      price_max: queryParams.price_max,
+      price_min: queryParams.price_min,
+      rating_filter: queryParams.rating_filter,
+      sort_by: queryParams.sort_by,
+      category: queryParams.category
+    },
+    isUndefined
+  )
+  return queryConfig
+}
