+++ src/pages/ProductList/ProductList.tsx
@@ -2,34 +2,14 @@ import { useQuery } from '@tanstack/react-query'
 import { AsideFilter } from './components/AsideFilter'
 import Product from './components/Product/Product'
 import { SortProductList } from './components/SortProductList'
-import useQueryParams from 'src/hooks/useQueryParams'
 import productApi from 'src/apis/product.api'
 import { Pagination } from 'src/components/Pagination'
 import { ProductListConfig } from 'src/types/product.type'
-import { omitBy, isUndefined } from 'lodash'
 import categoryApi from 'src/apis/category.api'
-
-export type QueryConfig = {
-  [key in keyof ProductListConfig]: string
-}
+import useQueryConfig from 'src/hooks/useQueryConfig'
 
 function ProductList() {
-  const queryParams = useQueryParams()
-  const queryConfig: QueryConfig = omitBy(
-    {
-      page: queryParams.page || '1',
-      limit: queryParams.limit || '5',
-      exclude: queryParams.exclude,
-      name: queryParams.name,
-      order: queryParams.order,
-      price_max: queryParams.price_max,
-      price_min: queryParams.price_min,
-      rating_filter: queryParams.rating_filter,
-      sort_by: queryParams.sort_by,
-      category: queryParams.category
-    },
-    isUndefined
-  )
+  const queryConfig = useQueryConfig()
   // Get Products
   const { data: productsData } = useQuery({
     queryKey: ['products', queryConfig],
