+++ src/pages/ProductList/components/SortProductList/SortProductList.tsx
@@ -1,11 +1,11 @@
 import { ChevronLeft, ChevronRight } from 'src/icons'
-import { QueryConfig } from '../ProductList'
 import { sortBy, order as orderConstant } from 'src/constants/product'
 import classNames from 'classnames'
 import { ProductListConfig } from 'src/types/product.type'
 import { Link, createSearchParams, useNavigate } from 'react-router-dom'
 import path from 'src/constants/path'
 import { omit } from 'lodash'
+import { QueryConfig } from 'src/hooks/useQueryConfig'
 interface Props {
   queryConfig: QueryConfig
   pageSize: number
